requirements.txt
# Deep Learning and Image Processing
tensorflow==2.18.0
numpy==2.0.2
Pillow==11.0.0
opencv-python==4.10.0.84
joblib==1.4.2

# Web Application
streamlit==1.40.2
scikit-learn==1.5.2
pandas==2.2.3
matplotlib==3.9.2

# Dependencies these packages need
scipy==1.14.1
protobuf>=3.20.0
typing-extensions>=4.12.2
pyarrow>=7.0.0